fix(payments-next,ui): Move PayPalScriptProvider to PaymentSection
This pull request

- updates `<Providers>` component by removing `<PayPalScriptProvider>`
- moves `<PayPalScriptProvider>` to `<PaymentSection>` component

Closes: FXA-11247
